chore: project folder structure reorganized
- now all our entities should be added into the database folder were they belong
