--- src/api/locations/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './location.entity';
