--- src/api/locations/index.ts
@@ -1,1 +1,0 @@
-export * from './entities';
