--- src/api/riders/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './rider.entity';
