--- src/api/riders/index.ts
@@ -1,2 +1,0 @@
-export * from './entities';
-export * from './rider.module';
