+++ src/database/entities/base.entity.ts
