+++ src/database/entities/index.ts
@@ -1,0 +1,3 @@
+export * from './base.entity';
+export * from './location.entity';
+export * from './rider.entity';
