+++ src/database/entities/location.entity.ts
@@ -1,6 +1,6 @@
 import { Entity, Column } from 'typeorm';
 
-import { Base } from '../../common';
+import { Base } from './base.entity';
 
 @Entity()
 export class Location extends Base {
